Parser: [emphasis] remove two-asterisk markdown version
First, this was planned so that it would be possible to mark emphasis
inside other emphasis.

Ideally, this would work like `*foo **bar***`; however, prohibiting the
inverse `**foo *bar***` would be annoying (editing that removed the
outer emphasis, would require to update all inner emphasis) and hard to
do (for it to give meaningful error messages, it would require us to
hold and pass around even more state).

On the other hand, implementing the inverse is also hard, since the
ending `***`is ambiguous; first, it would require us mutate tokens or
make something C-like where the lexer depends too on the parser; second,
we would need to be able to write a proper grammar for that, and while
the "longest-match" strategy is reasonable enough to solve lexer
ambiguities, it can mess things up in the parser.

Thus, I've removed second level emphasis with markdown syntax; this
might still be possible with commands.