Gen: [parse] change comment syntax from /**/,// to \''\
This makes it easier and safer to type slashes, paths and addresses, or
simply reduces the list of reserved characters by one.

An added benefit is that now it's easier to use the universal / as a
path separator, than the windows only \ (because the latter has to be
escaped).

Alternatives considered were:

```
7	?2		foo \' block '\ bar  \' line '\		used to be based on troff "
4	D (mod)		foo \+ block +\ bar  ++ line
5	Curl		foo |# block #| bar  || line
3	Haskell		foo {- block -} bar  {- line -}
9	?4		foo {* block *} bar  {* line *}		based on OCaml/Haskell
10	?5		foo |+ block +| bar  || line		based on C++|D
6	?1		foo |* block *| bar  || line		based on C++
8	?3		foo %* block *% bar  %% line		based on TeX
2	OCaml		foo (* block *) bar  (* line *)
1	C++		foo /* block */ bar  // line
```