Gen: [null] some work on making the parser Ast null safe
Explored changing from potentially Null<T> VElem and HElem to null safe
elements with explict 'empty` constructors for their corresponding
?Defs.

This can work (and should improve things), but it's not yet clear to me
how what should be the corresponding `pos` field for an empty
VEmpty/HEmpty definition.  Neither is how to actually keep track of
with, with our need for a token buffer.

In fact, this implementation is absurdly broken in respect to the
positions and the token buffer...