Gen: parser errors as ADTs and add `PositionTools.highlight`
This simplifies the flow of information by dropping the dependency to
the lexer in the parser errors (fixes #60).

By itself this would make the errors less informative, since we wouldn't
have enough access to the original source (or we would depend too
much on them being actual accessible files on the disk).  But, we
instead improved the human readable error output by adding a fancy
(although still experimental) `highlight(?lineLength)` api to the
`Position` tools.

Errors now look, to the regular user, something like this:

```
ERROR: Unknown command \facilisis; did you perhaps mean `\section`?  (sorry if the suggestion makes no sense)
  at guide/the-guide.src, line=14, columns=(347 to 356):
    ...eros interdum maximus. Fusce varius sodales risus, quis \facilisis turpis consec...
                                                               ^^^^^^^^^^
```

This still requires some fine tuning, as well as tests:

 - infer maximum line length automatically (most times), choose a
   default with more care or, maybe, let it as a advanced option (via
   environment variables)
 - add unit tests
 - possible revert 8720c82, since we're not using `Token.src` anywhere
   after all

The idea is to add this for the other types of errors as well!