manu: move the table of contents in mobile clients to the bottom
First, having it collapsed at the top was not working so well,
apparently.  Renato got lost and I also noticed that it was very
painfull to navigate.

However, the previous solution – to have it expanded at the _top_ – only
helped with one particular use case: browsing.  On the other hand, it
got in the way of the actual content of the page and would probably make
the users even more lost.

Thus, this patch experiments with having it expanded at the _bottom_ of
the page.  This works in conjuction with a new "/ ..." breadcrumb item
at the top.

Finally, this makes no changes to desktop clients (or, more specificaly,
clients with wide viewports).