Specify the canonical URL [manu]
Since the generated webpages are supposed to be accessed and indexed
from `/foo` and not `/foo.html`, this is a basic step to avoid undesired
links in search engines.

It's recommended that the webserver used is configured to set 302 (temp)
redirects from `/foo.html` to `/foo`, even if we can use it yet on this
repository (because we have open pull requests that would fail to load the
nav bar with this settings).