Extract manu into a separate tree
While manu was developed as part of the Online BRT Planning Guide, it
was always inteded to be general tool.

The commit history was extracted using `git subtree` and later cleaned
from LFS objects using `git lfs migrate`.

    git subtree split --prefix generator
    git lfs migrate export --include 'docs/**/*' --skip-fetch

The history starts shortly after the last proof-of-concept version,[1]
and includes up to the latest commit in the original repository.[2]

[1] Organize: separate generator folder (Apr 18 2016): https://github.com/ITDP/the-online-brt-planning-guide/commit/fe89ac4ff3152fa562ddcaa9c6f114bc4230a0e5
[2] Upgrade utf-8-validate for compatibility with Node v10 (Aug 27 2018): https://github.com/ITDP/the-online-brt-planning-guide/commit/9d3057044d8201a84d7c35fa67789e4cbcf099a4